segregate method visitors
