+++ classport-instr-agent/src/main/java/io/github/chains_project/classport/instrumentation/granularity/ClassInvocation.java
@@ -1,5 +1,9 @@
 package io.github.chains_project.classport.instrumentation.granularity;
 
+import io.github.chains_project.classport.commons.ClassportInfo;
+import org.objectweb.asm.MethodVisitor;
+import org.objectweb.asm.Opcodes;
+
 import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileWriter;
@@ -49,5 +53,37 @@ public class ClassInvocation implements RecordingStrategy {
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
+	}
+
+	@Override
+	public MethodVisitor startVisitor(MethodVisitor mv, String methodName, String className, ClassportInfo ann) {
+		// No specific visitor logic for class-level invocation
+		return new ClassInterceptor(mv, methodName, className, ann);
+
+	}
+}
+
+class ClassInterceptor extends MethodVisitor {
+	private final String methodName;
+	private final String className;
+	private final ClassportInfo ann;
+
+	public ClassInterceptor(MethodVisitor mv, String methodName, String className, ClassportInfo ann) {
+		super(Opcodes.ASM9, mv);
+		this.methodName = methodName;
+		this.className = className;
+		this.ann = ann;
+	}
+
+	@Override
+	public void visitCode() {
+		super.visitCode();
+		// Inject code to add to the queue every time the method is invoked
+		mv.visitLdcInsn(ann.group() + "," + ann.artefact() + "," + ann.version());
+		mv.visitMethodInsn(Opcodes.INVOKESTATIC,
+				"io/github/chains_project/classport/instrumentation/MethodInterceptorVisitor",
+				"addToInvokeLater",
+				"(Ljava/lang/String;)V",
+				false);
 	}
 }
